Add files via upload
An application through which a user can predict whether a person is suffering from heart disease or not by giving specific input. Techniques used for classifying are Naive Bayes, Logistic Regression. Datasets used in this project are Cleavland, Hungarian and VA long beach.The package used for training and testing is sklearn.metrics